requirements.txt
psycopg2-binary==2.9.9
Faker==20.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
pydantic==2.5.0
pandas==2.1.3
redis==5.0.1
python-dotenv==1.0.0
pydantic-settings==2.1.0